Code V1/requirements.txt
# Core AI/ML
torch
torchao
torchvision
transformers
ultralytics
moondream
einops
accelerate

# Image Processing
opencv-python
pillow

# Hardware (Raspberry Pi Only - Comment out on Windows)
RPi.GPIO
gpiozero
